Fix route and successfully send response
